# desktop/src/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aether",
    "description": "something something app",
    "contact": {
      "name": "you"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/goals": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "Get all goals",
        "operationId": "get_goals",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of goals per page (max 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Cursor for pagination",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of goals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGoals"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Create a new goal",
        "operationId": "create_goal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created goal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/goals/{goalId}/instances": {
      "get": {
        "tags": [
          "GoalInstances"
        ],
        "summary": "Get goal instances for a goal",
        "operationId": "get_goal_instances",
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of instances per page (max 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Cursor for pagination",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of goal instances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGoalInstances"
                }
              }
            }
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/goals/{goalId}/instances/current": {
      "get": {
        "tags": [
          "GoalInstances"
        ],
        "summary": "Get or create current goal instance",
        "operationId": "get_current_goal_instance",
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current goal instance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalInstance"
                }
              }
            }
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/goals/{id}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "Get goal by ID",
        "operationId": "get_goal_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goal found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Goals"
        ],
        "summary": "Update a goal",
        "operationId": "update_goal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated goal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Goal not found"
          },
          "409": {
            "description": "Conflict: Goal was modified by another device"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Goals"
        ],
        "summary": "Delete a goal",
        "operationId": "delete_goal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Goal deleted"
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/goals/{id}/tags": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Add tags to a goal",
        "operationId": "add_tags_to_goal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tags added to goal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "404": {
            "description": "Goal or tag not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Goals"
        ],
        "summary": "Remove tags from a goal",
        "operationId": "remove_tags_from_goal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Goal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tags removed from goal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          },
          "404": {
            "description": "Goal not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGoalRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "isNonRecurring": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "recurrenceAnchor": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recurrenceMeta": {
            "nullable": true
          },
          "recurrenceType": {
            "type": "string",
            "nullable": true
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Goal": {
        "type": "object",
        "description": "Goal model",
        "required": [
          "id",
          "name",
          "isNonRecurring",
          "timezone",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "isNonRecurring": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "recurrenceAnchor": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recurrenceMeta": {
            "nullable": true
          },
          "recurrenceType": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GoalInstance": {
        "type": "object",
        "description": "GoalInstance model",
        "required": [
          "id",
          "goalId",
          "periodStart",
          "status",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "goalId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "PaginatedGoalInstances": {
        "type": "object",
        "description": "Pagination response wrapper optimized for infinite scroll",
        "required": [
          "items",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean",
            "description": "Convenience flag for infinite scroll"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalInstance"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "Next page cursor (null when no more pages)",
            "nullable": true
          }
        }
      },
      "PaginatedGoals": {
        "type": "object",
        "description": "Pagination response wrapper optimized for infinite scroll",
        "required": [
          "items",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean",
            "description": "Convenience flag for infinite scroll"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Goal"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "Next page cursor (null when no more pages)",
            "nullable": true
          }
        }
      },
...
      "UpdateGoalRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "isNonRecurring": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "recurrenceAnchor": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recurrenceMeta": {
            "nullable": true
          },
          "recurrenceType": {
            "type": "string",
            "nullable": true
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}